# docs/openapi/control-plane.yaml
openapi: 3.1.0
info:
  title: nexus control-plane API
  version: "1.0.0"
  description: |
    Routing-plane administration surface. Authors tenancy and routing: accounts,
    workspaces, memberships, per-workspace auth-routes, and custom domains.

    Fail-closed: every data route requires a bearer credential that resolves to
    an INDIVIDUALLY IDENTIFIABLE admin token (`routing.admin_tokens`, minted via
    `POST /admin-tokens`; `ADMIN_TOKEN_PEPPER` keys verification). During the
    token migration only, the legacy shared `CONTROL_AUTH_TOKEN` is accepted iff
    `ADMIN_LEGACY_TOKEN_OK=true` (attributed `legacy-shared`, deprecation-warned).
    The process refuses to start without an explicit auth choice
    (`CONTROL_AUTH_DISABLED=true` is the dev opt-out). There is no workspace
    header. Every success body carries `"result":"ok"`.

    Admin action audit (admin-action-audit): every mutating call records one
    audit event ATOMICALLY with the mutation — an unrecorded mutation does not
    commit — and rejected authentications are recorded as denial events. Every
    mutating endpoint accepts the optional `x-acting-operator` header (see the
    ActingOperator parameter): recorded verbatim, marked asserted, never used
    for authorization. The ledger is queryable at `/audit/events` and
    exportable at `/audit/events/export`.

    Ids are SERVER-MINTED (`acct_<uuidv7>` / `ws_<uuidv7>`): creation returns the id
...
    `created:false` instead of minting a duplicate.

...
servers:
...
security:
  - bearerAuth: []
...
paths:
  /accounts:
    post:
      tags: [accounts]
      summary: Provision an account
      operationId: provisionAccount
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountBody' }
      responses:
        '200':
          description: Provisioned (idempotent)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountUpsertResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /accounts/{id}:
    get:
      tags: [accounts]
      summary: Read an account and its members
      operationId: getAccount
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: Account with members
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountRead' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: Unknown account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotFoundAccount' }
  /workspaces:
    post:
      tags: [workspaces]
      summary: Create a workspace (server-minted id)
      description: Create-only — never overwrites. The minted `ws_…` id is in the response.
      operationId: createWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkspaceCreateBody' }
      responses:
        '200':
          description: Created (`created:true`) or idempotency-key replay (`created:false`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkspaceCreateResult' }
        '400':
          description: Invalid target_pool (not in the allow-list), or invalid_idempotency_key
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: Unknown owning account (when account_id is supplied)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /workspaces/{id}:
    put:
      tags: [workspaces]
      summary: Reconfigure a workspace (plan/pool/features)
      description: Update-only — an unknown id is a 404, never an implicit create.
      operationId: reconfigureWorkspace
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkspaceReconfigureBody' }
      responses:
        '200':
          description: Reconfigured
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkspaceUpsertResult' }
        '400':
          description: Invalid target_pool (not in the allow-list)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: unknown_workspace
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    get:
      tags: [workspaces]
      summary: Read a workspace config
      operationId: getWorkspace
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Workspace config
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkspaceConfig' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: Unknown workspace
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotFoundWorkspace' }
  /workspaces/{id}/transfer:
    post:
      tags: [workspaces]
      summary: Transfer a workspace to another account
      operationId: transferWorkspace
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransferBody' }
      responses:
        '200':
          description: Transferred
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransferResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: Unknown target account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /workspaces/{id}/members:
    get:
      tags: [members]
      summary: List workspace memberships
      operationId: listMemberships
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Memberships
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MembershipList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [members]
      summary: Upsert a membership
      operationId: upsertMembership
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MembershipBody' }
      responses:
        '200':
          description: Upserted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MembershipUpsertResult' }
        '400':
          description: Invalid member_type (not staff/customer)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: Unknown workspace
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /workspaces/{id}/members/{sub}:
    delete:
      tags: [members]
      summary: Remove a membership
      operationId: deleteMembership
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - name: sub
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Removed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MembershipUpsertResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /workspaces/{id}/auth-routes:
    get:
      tags: [auth-routes]
      summary: List auth-route rules
      operationId: listAuthRoutes
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Rules
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthRouteList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [auth-routes]
      summary: Upsert an auth-route rule
      operationId: upsertAuthRoute
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthRouteBody' }
      responses:
        '200':
          description: Stored rule (echoed)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthRouteUpsertResult' }
        '400':
          description: invalid_prefix, or requirements set with auth_required=false (requirements_need_auth)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: Unknown workspace
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [auth-routes]
      summary: Delete an auth-route rule
      operationId: deleteAuthRoute
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthRouteDelete' }
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthRouteDeleteResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /domains:
    post:
      tags: [domains]
      summary: Upsert a domain
      description: Always created verified=false; use declare→verify to make it route.
      operationId: upsertDomain
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DomainBody' }
      responses:
        '200':
          description: Upserted (unverified)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DomainUpsertResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /domains/declare:
    post:
      tags: [domains]
      summary: Declare a custom domain and start verification
      operationId: declareDomain
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeclareBody' }
      responses:
        '200':
          description: Declared; returns the DNS proof record to publish (absent if already verified)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeclareResult' }
        '400':
          description: invalid_domain
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402':
          description: quota_exceeded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuotaExceeded' }
        '404':
          description: unknown_workspace
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: domain_taken
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /domains/{domain}/verify:
    post:
      tags: [domains]
      summary: Trigger DNS-ownership verification
      description: No request body.
      operationId: verifyDomain
      parameters:
        - $ref: '#/components/parameters/Domain'
      responses:
        '200':
          description: Verified
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerifyResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: no_challenge
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '410':
          description: challenge_expired
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422':
          description: proof_not_found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503':
          description: resolution_failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /domains/{domain}:
    delete:
      tags: [domains]
      summary: Delete a domain
      operationId: deleteDomain
      parameters:
        - $ref: '#/components/parameters/Domain'
        - name: wildcard
          in: query
          required: false
          description: Selects which row to drop; does not change the response shape.
          schema: { type: boolean }
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DomainDeleteResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /audit/events:
    get:
      tags: [audit]
      summary: Query the admin audit ledger
      description: |
        Read-only, time-ordered (by `event_id` — `aev_<uuidv7>` sorts by event
        time), cursor-paginated. Filters compose (AND). No mutation endpoint
        over events exists anywhere (append-only ledger, admin-action-audit).
      operationId: listAuditEvents
      parameters:
        - { name: from, in: query, required: false, description: "Inclusive lower bound on occurred_at (RFC 3339)", schema: { type: string } }
        - { name: to, in: query, required: false, description: "Exclusive upper bound on occurred_at (RFC 3339)", schema: { type: string } }
        - { name: actor, in: query, required: false, description: "Exact acting credential id (or a reserved actor id)", schema: { type: string } }
        - { name: target, in: query, required: false, description: "Exact target resource id", schema: { type: string } }
        - { name: cursor, in: query, required: false, description: "Resume strictly after this event_id (previous page's next_cursor)", schema: { type: string } }
        - { name: limit, in: query, required: false, description: "Page size (clamped to 1..=1000; default 100)", schema: { type: integer } }
      responses:
        '200':
          description: One page of events; an empty `events` array means done.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuditEventsPage' }
        '400':
          description: Malformed from/to bound
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /audit/events/export:
    get:
      tags: [audit]
      summary: Export audit events as NDJSON
      description: |
        Streams one JSON event per line for the time range (machine-readable
        evidence for external audit consumers). Read-only — the ledger is
        unchanged.
      operationId: exportAuditEvents
      parameters:
        - { name: from, in: query, required: false, description: "Inclusive lower bound on occurred_at (RFC 3339)", schema: { type: string } }
        - { name: to, in: query, required: false, description: "Exclusive upper bound on occurred_at (RFC 3339)", schema: { type: string } }
      responses:
        '200':
          description: NDJSON stream of AuditEvent lines
          content:
            application/x-ndjson:
              schema: { type: string }
        '400':
          description: Malformed from/to bound
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /admin-tokens:
    post:
      tags: [admin-tokens]
      summary: Issue a named admin token
      description: |
        Mints an individually identifiable admin credential for ONE caller
        (attribution, not authorization). The plaintext secret is returned ONCE
        and never persisted — only its peppered HMAC lands in
        `routing.admin_tokens`. Requires ADMIN_TOKEN_PEPPER (else 503). The
        issuance itself is audited (`admin_token.issue`). First-token
        chicken-and-egg: mint it while the legacy shared token
        (ADMIN_LEGACY_TOKEN_OK=true) still authenticates this call.
      operationId: issueAdminToken
      parameters:
        - { $ref: '#/components/parameters/ActingOperator' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IssueAdminTokenBody' }
      responses:
        '201':
          description: Token issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IssuedAdminToken' }
        '400':
          description: Missing/empty name
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/AdminTokenMgmtUnconfigured' }
  /admin-tokens/{id}/rotate:
    post:
      tags: [admin-tokens]
      summary: Rotate a named admin token
      description: |
        Mints a NEW secret under the same name with lineage (`rotated_from`) and
        revokes the old credential — other callers' tokens are untouched.
        Audited (`admin_token.rotate`). No request body.
      operationId: rotateAdminToken
      parameters:
        - { $ref: '#/components/parameters/AdminTokenId' }
        - { $ref: '#/components/parameters/ActingOperator' }
      responses:
        '201':
          description: Token rotated; the NEW token_id + one-time secret.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IssuedAdminToken' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: No active token with that id
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503': { $ref: '#/components/responses/AdminTokenMgmtUnconfigured' }
  /admin-tokens/{id}/revoke:
    post:
      tags: [admin-tokens]
      summary: Revoke a named admin token
      description: |
        Status flip; the credential is rejected on subsequent calls while every
        other caller's keeps working. Idempotent (`revoked:false` when already
        revoked/unknown). Audited (`admin_token.revoke`). No request body.
      operationId: revokeAdminToken
      parameters:
        - { $ref: '#/components/parameters/AdminTokenId' }
        - { $ref: '#/components/parameters/ActingOperator' }
      responses:
        '200':
          description: Revoked (idempotent)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminTokenRevokeResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/AdminTokenMgmtUnconfigured' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: CONTROL_AUTH_TOKEN
  parameters:
    AccountId:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token (a denial event is recorded)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    AccountBody:
      type: object
      description: No `account_id` — the id is server-minted and returned in the response.
      properties:
        owner_sub:
          type: string
          description: Becomes the account's first (owner) member.
        name:
          type: string
          default: ""
          description: Display label only — no identity or uniqueness semantics.
        payer_ref: { type: [string, "null"] }
        idempotency_key:
          type: [string, "null"]
          maxLength: 128
          description: Optional replay guard (opaque to nexus, e.g. `signup:<sub>`).
      required: [owner_sub]
      additionalProperties: false
    AccountUpsertResult:
      type: object
      properties:
        result: { type: string, const: ok }
        account_id: { type: string }
        created:
          type: boolean
          description: false = an idempotency-key replay returned the original account.
      required: [result, account_id, created]
...
    AccountRead:
      type: object
      properties:
        account: { $ref: '#/components/schemas/Account' }
        members:
          type: array
          items: { $ref: '#/components/schemas/AccountMember' }
      required: [account, members]
...